Rename playlist.js to playlistPage.js
